improve the document about contributing to bug reports
Recently we hit many correctness bugs, and some of them do not get enough attention. I want to highlight the importance of data correctness/data loss issues, and suggest contributors to describe how the bug was introduced, so that people can quickly understand the bug. It's also possible that the reporter doesn't realize it's a correctness bug, by looking at how the bug was introduced, other people can point it out.

Author: Wenchen Fan <wenchen@databricks.com>

Closes #153 from cloud-fan/contribute.