Change the URL "www.apache.org/dist" to "downloads.apache.org" as structural change on ASF INFRA
Recently, ASF INFRA team has sent an email "[NOTICE] Structural changes to Apache downloads" to encourage ASF projects to change links from `www.apache.org/dist/` to `downloads.apache.org/`.

According to the mail content it's not a thing to be hurry as the old URL would be redirected with new URL, but as the old URL will be deprecated soon, it would be better to handle it early and forget about it.

This patch only changes URLs which is a straightforward change - so I simply skipped building content - please let me know if we would want to build and double-check.

Author: Jungtaek Lim (HeartSaVioR) <kabhwan.opensource@gmail.com>

Closes #262 from HeartSaVioR/change-urls-apache-dist-to-download-apache.