Add Data Mechanics to Powered By
Data Mechanics is a managed Spark platform that can be deployed on a Kubernetes cluster inside our customers cloud accounts. We'd love to be on the Powered By Spark page (along other Spark platforms).

We contribute to open source projets in the Spark ecosystem (Spark on Kubernetes operator, Data Mechanics Delight). We also use Spark internally for our recommendation engine and logs processing.

I tried to be objective / avoid marketing in the description, but I'm open to feedback on changing it. Thanks!

Author: Jean-Yves Stephan <jean-yves@datamechanics.co>

Closes #299 from jystephan/datamechanics-poweredby.