Lock Jekyll version to 3.6.3
Pin Jekyll to version 3.6.3.

As this version is where the diff is minimal, checked by the script:
```
#!/bin/zsh

versions=('3.7.3' '3.7.2' '3.7.0' '3.6.3' '3.6.2' '3.6.1' '3.6.0' '3.5.2'
'3.5.1' '3.5.0' '3.4.5' '3.4.4' '3.4.3' '3.4.2' '3.4.1' '3.4.0')

for i in $versions; do
  gem uninstall -a -x jekyll rouge
  gem install jekyll --version $i
  jekyll build
  git diff --stat
  git reset --hard HEAD
done
```

The first commit contains those minimal changes which was by running `jekyll build` with version 3.6.3:
- reordering in the sitemap.xml
- some extra space at the end of line in community.html

The 2nd commit updates the README.md and introduces the Gemfile to pin Jekyll to 3.6.3.

Author: attilapiros <piros.attila.zsolt@gmail.com>

Closes #303 from attilapiros/fix_jekyll_version.