Fix confict warning on spark-summit-east-agenda-posted
When we use `jekyll build` to generate the htmls the below error is raise:
```
Conflict: The following destination is shared by multiple files.
The written file may end up with unexpected contents.
/Users/jiangyikun/spark/spark-website/site/news/spark-summit-east-agenda-posted.html
 - spark-website/news/_posts/2015-01-21-spark-summit-east-agenda-posted.md
 - spark-website/news/_posts/2016-01-14-spark-summit-east-agenda-posted.md
```
The reason is there were two files have same suffix:
- 2015-01-21-spark-summit-east-agenda-posted
- 2016-01-14-spark-summit-east-agenda-posted

And there ware only one page is generated, the contents is overwrite by the
page with 2016-01-14 prefix, and the 2015-01-21 is lost. See [1].

This patch adds diff suffix for 2015-01-21-spark-summit-east-agenda-posted.md
and 2016-01-14-spark-summit-east-agenda-posted.md to fix this confict issue.

[1] http://spark.apache.org/news/spark-summit-east-agenda-posted.html