fix display problem
<!-- *Make sure that you generate site HTML with `bundle exec jekyll build`, and include the changes to the HTML in your pull request. See README.md for more information.* -->
I think this `|` inside ` Support (IGNORE | RESPECT) ...` caused display problem. The dot looks strange and the `|` in between of `IGNORE` and `RESPECT` is gone.
```
- [[SPARK-30789]](https://issues.apache.org/jira/browse/SPARK-30789): Support (IGNORE | RESPECT) NULLS for LEAD/LAG/NTH_VALUE/FIRST_VALUE/LAST_VALUE
```
<img width="891" alt="Screen Shot 2022-01-28 at 2 39 37 PM" src="https://user-images.githubusercontent.com/13592258/151631202-b3219f5e-9c88-4cb9-a60d-bd2de2536f1d.png">

I will also fix the spaces in this file. All the other release md files have two spaces before each of the line for the jira list, but my intelliJ setting is not right, all the spaces are gone so I will manually add these spaces back.

Author: huaxingao <huaxin.gao11@gmail.com>

Closes #377 from huaxingao/fixspace.