Add Matomo analytics
We removed Google Analytics last year, to comply with ASF privacy policies.
There was/is an option to use an ASF-hosted analytics service called Matomo.
I've had a request for that, so propose we try adding this to the website and see how we like it, then consider putting it in release docs.

See https://github.com/apache/spark-website/pull/384

Author: Sean Owen <srowen@gmail.com>

Closes #479 from srowen/Matomo.