update the canonical link, due to a change in some addresses in the latest version of the document
The pr is followup https://github.com/apache/spark-website/pull/482.

https://github.com/apache/spark-website/pull/482#issuecomment-1765322679
As discussed above, due to changes in some document addresses after version `3.3.0`, `the canonical link` is incorrect. We are now correcting it.

Base on rule:
For file that do not exist in version `latest`, delete the corresponding ref and update it for the rest.

Author: panbingkun <pbk1982@gmail.com>

Closes #483 from panbingkun/canonical_links_followup.