Guide release manager to give a last chance to backup docs artifact before removing it
In release process page, we guide the release manager "too lately" about keeping the generated docs of the latest RC. The guidance sentence appears later than removing docs from svn.

https://spark.apache.org/release-process.html - read through the section "finalize the release".

If release manager does not have any backup, they are expected to build the docs from scratch, and it's really painful to set up languages and corresponding libraries locally to build the doc (release script with docker is no longer usable at this time). It would be better to remind that they can take a backup before removing the artifact, which they will use the backup very soon.

Here is the screenshot of the change - bold NOTE sentence:

<img width="994" alt="스크린샷 2024-02-26 오후 4 33 07" src="https://github.com/apache/spark-website/assets/1317309/949ab008-bb84-4eaf-880f-084ebca3bd90">

Author: Jungtaek Lim <kabhwan.opensource@gmail.com>

Closes #505 from HeartSaVioR/guide-do-backup-docs-artifact-on-release-finalization.